requirements.txt
autopep8==2.0.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
protobuf==3.20.3
pycodestyle==2.10.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.12.0
